requirements.txt
aiogram==3.11.0
aiohttp==3.10.10
python-dotenv==1.0.1
feedparser==6.0.10
trafilatura==1.13.1